+++ lib/model.dart
@@ -12,6 +12,7 @@ import 'package:flutter/material.dart';
 import 'dart:async';
 import 'package:rxdart/rxdart.dart';
 import 'dart:typed_data';
+import 'package:json_annotation/json_annotation.dart';
 
 // Importing the UI library from Research Package
 import 'package:research_package/ui.dart';
@@ -49,3 +50,6 @@ part 'src/model/Abstracts.dart';
 // BLoCs
 part 'src/model/blocs/BlocTask.dart';
 part 'src/model/blocs/BlocQuestion.dart';
+
+// JSON
+part 'model.g.dart';