+++ lib/src/model/answerformat/RPIntegerAnswerFormat.dart
@@ -2,11 +2,14 @@ part of research_package_model;
 
 // TODO: not implemented yet
 /// Not fully implemented yet
+@JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
 class RPIntegerAnswerFormat extends RPAnswerFormat {
   int _minValue;
   int _maxValue;
 
-  RPIntegerAnswerFormat(this._minValue, this._maxValue);
+  RPIntegerAnswerFormat();
+
+  RPIntegerAnswerFormat.withParams(this._minValue, this._maxValue);
 
   @override
   get questionType {
@@ -17,5 +20,16 @@ class RPIntegerAnswerFormat extends RPAnswerFormat {
 
   int get minValue => _minValue;
 
+  set maxValue(int maxValue) {
+    this._maxValue = maxValue;
+  }
+
+  set minValue(int minValue) {
+    this._minValue = minValue;
+  }
+
   //TODO: Validation?
+
+  factory RPIntegerAnswerFormat.fromJson(Map<String, dynamic> json) => _$RPIntegerAnswerFormatFromJson(json);
+  Map<String, dynamic> toJson() => _$RPIntegerAnswerFormatToJson(this);
 }