+++ CHANGELOG.md
@@ -1,3 +1,17 @@
-## [0.0.1] - TODO: Add release date.
+## 0.0.2
+ * added initial support for serialization to/from JSON
+ * JSON serialization is available for these classes:
+    * `RPAnswerFormat`
+    * `RPChoiceAnswerFormat`
+    * `RPIntegerAnswerFormat`
+    * `RPConsentDocument`
+    * `RPConsentSection`
+    * `RPSignatureResult`
+    * `RPStepResult`
+    * `RPTaskResult`
+    * `RPChoice`
 
-* TODO: Describe initial release.
+## 0.0.1
+ * initial release
+ * entire framework done
+ * support for SingleChoice question type